requirements.txt
# The Chocolate Factory - Mini-MRP
# Python 3.10+

fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
jinja2>=3.1.3
python-multipart>=0.0.6
